+++ src/SubChunk1.java
@@ -1,11 +1,11 @@
 public class SubChunk1 {
-    public byte[] subChunk1Id;
-    public byte[] subChunkSize;
-    public byte[] audioFormat;
-    public byte[] numChannels;
-    public byte[] sampleRate;
-    public byte[] byteRate;
-    public byte[] blockAlign;
-    public byte[] bitsPerSample;
+    public int[] subChunk1Id;
+    public int[] subChunkSize;
+    public int[] audioFormat;
+    public int[] numChannels;
+    public int[] sampleRate;
+    public int[] byteRate;
+    public int[] blockAlign;
+    public int[] bitsPerSample;
 
 }
