+++ src/SubChunk2.java
@@ -1,5 +1,5 @@
 public class SubChunk2 {
-    public byte[] subChunk2Id;
-    public byte[] subChunk2Size;
-    public byte[] data;
+    public int[] subChunk2Id;
+    public int[] subChunk2Size;
+    public int[] data;
 }
